Extract UIView generic rendering logic
